pyproject.toml
[project]
name = "wiki-agent"
version = "0.1.0"
description = "A agent to generate repo markdown automatically"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.9.0",
    "PyGithub>=2.0.0",
    "requests>=2.32.0",
    "tqdm>=4.67.1",
    "langgraph>=0.2.0",
    "langchain-ollama>=0.2.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.12",
    "langchain-deepseek>=0.1.4",
    "tree-sitter>=0.24.0",
    "tree-sitter-cpp>=0.23.4",
    "tree-sitter-python>=0.25.0",
    "tree-sitter-javascript>=0.25.0",
    "tree-sitter-go>=0.25.0",
    "tree-sitter-java>=0.23.5",
    "tree-sitter-rust>=0.24.0",
    "tree-sitter-c>=0.24.1",
    "langchain-community>=0.3.31",
    "langchain-huggingface>=0.3.1",
    "sentence-transformers>=5.1.2",
    "langchain-chroma>=0.2.0",
    "unstructured[md]>=0.18.15",
    # backend
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.hatch.build.targets.wheel]
packages = ["agents", "app", "utils"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
